wip fixing tests
